refactor(video-conf): one call row instead of three components
`CallListItem` held the layout, `OngoingCallRow` and `RingingCallItem` each
filled its two slots, and `OngoingCallsList` re-derived the ringing state to
choose between them. It is the same row either way — the sidebar's room item
with a video mark — differing only in what sits at its edges.

Now one component reads the state off the call: ringing puts "Ringing…" where
the time goes and offers silence beside decline; a call that can still be
declined gets the button; a declined one keeps its place and says so. The list
maps calls to rows and no longer branches.

11 files to 7, with every assertion and story kept — the two specs and two
story files merge into one each, grouped by the state they describe.
